[V63007] Fix triggers/compile subtest failure due to no triggers installed if compile fails (GTM-9047)
Previously, certain types of M code used to show a compilation error but the corresponding trigger
would still be installed. Now, as part of GTM-9047 changes (see relevant release note below), the
trigger no longer installs so the reference file is fixed to reflect the new output. The M program
is also slightly enhanced to make the reference file more readable.

Release note of GTM-9047 relevant to this commit

* ZTRIGGER() and MUPIP TRIGGER don't install trigger definitions with XECUTE strings that do not
  compile without error; previously they did.