[V63007] Fix v44003a/C9D12002478 subtest reference file due to additional compile time messages in case of MAXARGCNT error (GTM-9043)
This subtest uses too many concatenation operands (_) in one line and induces the MAXARGCNT error.

Below is the full release note of GTM-9043.

GT.M detects the case of more concatenation operands in a row than it can handle when parsing the
source code; previously it detected this at code generation, which meant it always failed to create
an object file in this case. (GTM-9043)