[V63007] Fix v62001/gtm8094pipproc subtest reference file due to V63007 code changes
The test sets white-box env vars to induce certain error code paths. Those paths print
the current "errno" global in the DEVOPENFAIL error message. This "errno" value changed
from ENO2 to UNKNOWN or ENO22. Most likely that is because of the code changes done in
V63007 so we come in to sr_unix/iopi_open.c with a different errno based on the test random
settings. In any case, since these are white-box code paths, the system "errno" is not
actually set to a valid error but instead we print whatever this global variable contains
at that point in time. Therefore the code is fixed separately to initialize "errno" to
a specific value in the white-box case (see sr_unix/iopi_open.c) that way we always have
the same error number in the test outputs and so can have a deterministic reference file
which is now updated.