[V63007] Fix v62001/gtm4414 subtest failure due to JOBLVN2LONG error no longer showing up in .mje (GTM-9058)
Previously, the test used to search for JOBLVN2LONG in the .mje file. This can no longer be done.
Instead we need to only verify this error shows up in the output of the M program issuing the JOB
command. This is now done in the test.

Below is the release note of GTM-9058.

GT.M handles JOB errors differently. Errors associated with the specified routine, label, or offset
appear in the error file of the JOBbed process in detail in addition to the JOBFAIL error received
by the original process. Previously GT.M did not report the more specific errors and did not start
the JOBbed process. GT.M detects and reports JOBLVN2LONG errors in the original process; GT.M does
not start the JOBbed process in this case. Previously the JOBbed process would report JOBLVN2LONG
to its error file. (GTM-9058)