[V63007] Fix v63002/gtm8694 and v63005/gtm8877 subtest failures due to new LIBRARY restriction facility (GTM-7318)
As part of GTM-7318, the release note mentions a new APD_ENABLE facility. But as part of the
code changes, an undocumented LIBRARY facility was also added. This causes any dlopen() calls
to be validated for any current restrictions and a RESTRICTEDOP error is newly issued. This
means when we have a scenario where all restrictions are enabled (e.g. when we only have
read-only access to $ydb_dist/restrict.txt), the LIBRARY restriction causes an error way before
any other restrictions are even checked so the reference file needs to be adjusted accordingly
with "dlopen()" entries. The dlopen() output varies depending on various randomizations in the
test system framework (e.g. libicuio.so if UTF-8 is enabled, libhugetlbfs.so if hugepages are
enabled, libposix.so if posix plugin is in $zroutines etc.). Therefore we have a *.so to account
for these various cases in the reference file.