Exclude manually_start/gds_max_blk subtest on hosts where env var ydb_test_exclude_gds_max_blk is set to 1
On a Ubuntu 16.04 box VM, we see the following messages whenever this test runs.

kernel: XFS: mupip(59828) possible memory allocation deadlock size 95840 in kmem_alloc (mode:0x2400240)

And suspect the issue to be related to the underlying virtualization.
So we disable this subtest on that particular host by setting the ydb_test_exclude_gds_max_blk
env var to 1 only on that host.