[V63007] Fix triggers/merrorhandling subtest failure due to GTM-9047 by avoiding compile time error
merrorhandling.m contains a trigger definition with -xecute="xecute ""set 1=1""".
This causes a compile-time error due to GTM-9047 and in turn changes the flow of the entire test.
This is now avoided by adding a QUIT (i.e. -xecute="xecute ""set 1=1"" quit") this way the
literal optimization of the XECUTE command does not happen at trigger install time letting the
test flow run just like it was before.

The reference file needed a minor adjustment due to this as well as due to a prior change to
merrorhandling.m to do a $ztrigger (instead of redirecting the $ztrigger output to a file and
filtering out just the needed lines) therefore resulting in a lot of new lines corresponding to
$ztrigger output now added.