[YDB#297] Fix r124/ydb297 subtest failure after integrating GT.M V6.3-006
In GT.M V6.3-006, the lock structures were reworked. This caused the hash values to
change so subscripts that used to hash to the same value no longer did. New subscripts
were determined to cause > 31 collisions with V6.3-006 and ydb297.m fixed accordingly.
We were able to only determine 47 collisions (instead of the previous 72 collisions).
This caused the collision counter to show up as 16 (instead of 41 previously). But it
is still a non-zero value which is what the test cares about.

In addition, the LOCKSPACEINFO messages had change in the format due to fewer number
of subscripts in use. The reference file has been fixed accordingly.