Fix rare v62000/gtm8086 subtest hang in JNLSWITCHRETRY test path (test of YDB#235)
As part of a prior commit 794ed6d0cbeb20f35efcb89d7a0180af9663fa76, a rare test hang of v62000/gtm8086
was fixed. In that, it was mentioned that other such hangs could still happen. We had one such hang
occur again.

In this case, the source server was stuck waiting for crit. Below is its C-stack.

 #0  pthread_mutex_timedlock () from /usr/lib64/libpthread.so.0
 #1  gtm_mutex_lock () at sr_unix/mutex.c:766
 #2  grab_crit () at sr_unix/grab_crit.c:114
 #3  repl_ctl_create () at sr_port/gtmsource_ctl_init.c:181
 #4  gtmsource_ctl_init () at sr_port/gtmsource_ctl_init.c:437
 #5  gtmsource_get_jnlrecs () at sr_unix/gtmsource_process_ops.c:976
 #6  gtmsource_process () at sr_unix/gtmsource_process.c:1547
 #7  gtmsource () at sr_unix/gtmsource.c:527
 #8  mupip_main () at sr_unix/mupip_main.c:123
 #9  dlopen_libyottadb () at sr_unix/dlopen_libyottadb.c:148
 #10 main () at sr_unix/mupip.c:19

While the process holding crit was waiting for the instance to be unfrozen while in jnl_file_close().
Below is its C-stack.

 #0  clock_nanosleep () from /usr/lib64/libc.so.6
 #1  wait_for_repl_inst_unfreeze_nocsa_jpl () at sr_port/anticipatory_freeze.h:503
 #2  wait_for_repl_inst_unfreeze () at sr_port/anticipatory_freeze.h:526
 #3  jnl_file_close () at sr_port/jnl_file_close.c:195
 #4  jnl_file_extend () at sr_unix/jnl_file_extend.c:204
 #5  t_end () at sr_port/t_end.c:1298
 #6  gvcst_put2 () at sr_port/gvcst_put.c:2654
 #7  gvcst_put () at sr_port/gvcst_put.c:299
 #8  op_gvput () at sr_port/op_gvput.c:77

In the prior commit hang case, there was a workaround (of bumping the journal autoswitchlimit).
But in this case, there is no such workaround.

So instead we take the approach of detecting if there is a hang while running the $MSR SYNC INST1 INST2
(inside RF_sync.csh) and if so do a MUPIP REPLIC -SOURCE -FREEZE=OFF to unfreeze the instance thereby
letting the crit holding process to continue and in turn letting the source server to continue clearing
the backlog. This is achieved using an env var that is set in gtm8086.csh before invoking $MSR SYNC.