Fix rare ideminter_rolrec/mupipstop_rollback_or_recover subtest failure due to white-box env vars not being set in case of a crash test
We had one case of the above subtest fail with the following core.

(gdb) where
 #0 __pthread_kill (threadid=<optimized out>, signo=3) at ../sysdeps/unix/sysv/linux/pthread_kill.c:57
 #1 gtm_dump_core () at sr_unix/gtm_dump_core.c:74
 #2 gtm_fork_n_core () at sr_unix/gtm_fork_n_core.c:163
 #3 ch_cond_core () at sr_unix/ch_cond_core.c:80
 #4 rts_error_va (csa=0x0, argcnt=7, var=...) at sr_unix/rts_error.c:192
 #5 rts_error_csa (csa=0x0, argcnt=7) at sr_unix/rts_error.c:99
 #6 wcs_flu (options=83) at sr_unix/wcs_flu.c:456
 #7 tp_tend () at sr_port/tp_tend.c:810
 #8 op_tcommit () at sr_port/op_tcommit.c:496
 #9 stkok3 () at sr_armv7l/opp_tcommit.s:38

(gdb) f 6
 #6 wcs_flu (options=83) at sr_unix/wcs_flu.c:456
 456  assert((WBTEST_CRASH_SHUTDOWN_EXPECTED == ydb_white_box_test_case_number) /* see wcs_phase2_commit_wait.c */

The assert failed because we are not able to flush the dirty buffers and the only situation we know of
this can happen is if processes were kill -9ed but the corresponding white-box env vars indicating a
crash were not set.

This test did actually choose to crash processes and does have logic to set the white-box env vars
but that logic was enabled only if online rollback was randomly chosen. This is because the prior
test failure where we had a similar assert fail was when online rollback was enabled.

But this time around, the test chose RECOVER and still failed. Therefore clearly the white-box env vars
should be set even for RECOVER (not just for ROLLBACK).

The code is now reworked to fix this. In addition, the variable was named inappropriately. It was
named "crash_gtm_updates" and when it was set to 0, the test chose crash and if it was 1 the test
chose a clean shutdown. This is now fixed so "crash_updates" (the variable name is now changed)
when set to 1 implies a crash shutdown and if 0 implies a clean shutdown.

Note that rolrec_intr_stop_idemp.csh is invoked by 3 subtests (idemp_rollback_or_recover,
interrupted_rollback_or_recover and mupipstop_rollback_or_recover). They call it with IDEMP/CRASH/STOP
respectively. In the CRASH case, the white-box env vars are already set so no changes needed. This
test failure (mupipstop_rollback_or_recover subtest) was using the STOP case. But looks like even
in the IDEMP case (i.e. idemp_rollback_or_recover subtest) we could encounter this issue.