Fix rare v63000/gtm6220 subtest failure if test runs within +10/-10 seconds of midnight
We had a test failure where the v63000/gtm6220 subtest with the following diff.

> %YDB-E-ZDATEBADTIME, $ZDATE() time argument -4 is less than 0 or greater than 86399 (the $HOROLOG value for a second before midnight)

The logic that created the files beforetime.txt and aftertime.txt in gtm6220.csh assumes
that the seconds value of $H is never < 10 or > 86400-10. That is exactly what happened
in this one test run and the calculation done in the test script caused the seconds
part of $H to go negative (subtracting 10 seconds from $h=xxxx,6 resulted in xxxx,-4
causing the ZDATEBADTIME error).

This is now fixed by enhancing the test framework M program difftime.m to have an entryref
addtime() that takes in a $h and number (in seconds) as input and adds the two and returns
a $h quantity. This is now used in gtm6220.csh.