[ydb#429] Fix of ydb429 to not erroneously include gtm_test_nopriorgtmver in it's env grep
Previously ydb429 would use `env` to print the environment and then grep
for the variables that should be set by ydb_env_set. This grep included
the string gtmver which would also print out gtm_test_nopriorgtmver. The
addition of ^re= to the regex fixes this. Added this to all lines
in case this issue comes up somewhere else.