[YottaDB/Lang/YDBGo#18] Fix v61000/setitimer_fail and v54003/C9K11003340 subtest failures since HANG no longer starts timer
For the v61000/setitimer_fail subtest, use a database update instead of a hang to start a timer
now that HANG does not start a timer after recent changes to YDB repo due to YottaDB/Lang/YDBGo#18.

For the v54003/C9K11003340 subtest, update the reference file to reflect new code behavior.