Revert "[#410] Add r126/ydb410 subtest to test signal forward/noforward."
This reverts commits e705a069bd13e9fa6494a0e10353b0e42f5dd29d and 7f496185145460984d47942952f4b29e3e59fbb6
(test addition and fix).

With this update, some regression tests with versions of Go past 1.10 failed with destroyed
stacks, which did not fail with this update. Although no cause-effect relationship could
reasonably be established, the correlation between passing and failing tests and this update
means that this update is held off, at least for the present. So it is reverted for now.