[YDB#478] Fix some inconsistencies in test that drops output or makes it appear out-of-order
This test has two goroutines that are both sending output to the reference file but were doing
so in a non-deterministic manner. These changes make the output order deterministic. Note that
Go output to stdout/stderr is NOT buffered by default so no flushes were added. We just make
sure to get rid of output that doesn't immediately precede a synchronization function of some
sort (like wg.Done()) or come after a call like wg.Wait().