Fix more repeatability issues by moving a message out of a goroutine and into the main
Since this test has both a main and a go routine (techincally 2 goroutines) and since
both do output to the reference file, it is important to make sure output is at given
"sync points" so the order of the output is absolutely predictable. When the goroutine
actually started running varied on some platforms resulting in changes in output order.
This change fixes that issue.