Fix test issues related to presence of gawk 5.0.1
Tthe files in this project were modified due to gawk 5.0.1 starting to show up on newer systems.
The new gawk raises a warning whenever a character is escaped that doesn't need to be escaped.
This was causing several tests to fail. Modifying these files to no longer escape those
characters fixed this issue.