Addressed simpleapi test failure
Sometimes, the simpleapi/externalcall test has been failing with the timer triggering after ydb_start_cancel instead of before. This change, which replaces a call to the deprecated usleep with a macro that calls nanosleep, has been tested 500 times on a machine where the test had failed and passed all 500 times.