Fix r122/tprestart failures due to no messages in log
On some platforms, this test could fail because too few tprestart messages were generated during the duration of the test. This change addresses that by retrying the tprestarts with a lower delta if no messages are generated for the log.

This approach was chosen instead of lowering the delta because:

1) It was possible to generate failures due to an empty log with a delta of 3, 4 or 5. This meant that the only viable deltas were 1 and 2 which would mean a dramatic increase in the number of tprestarts in the system log.
2) This change increases the average amount of time required to run the test by less than a quarter of a second per test on a test machine where it was periodically failing.