[YDB#503] Update existing tests to address test failures caused by previous changes
The changes for #503 addressed a bug where DRD was incorrectly being incremented when the database was set to the MM access method. However, some tests made the assumption that DRD was not zero and these tests would fail if run with the MM access.

When the test system picks BG as the access method, there will be no difference in the output. If the access method is MM, it will now expect DRD to be 0.