[YDB#518] Temporarily disable r130/ydb518 test on 32 bit machines
The new ydb_int64_t and ydb_uint64_t types for call-ins cause compile errors on 32 bit builds. As this could take a while to fix and we're approaching the release of r130, we decided to hold off on fixing this until after the release of r130 and thus are disabling the test for now on 32 bit machines. This should be undone when YDB#544 is addressed.