[YDB#349] Address r124/ydb349 test failures
The ydb349 subtest was failing if the random value for RESERVED_BYTES was between 4001 and 4012 because the maximum number of RESERVED_BYTES is 4000. This fixes the test failures by ensuring that the test generates a random value for RESERVED_BYTES between 1 and 4000. The maximum number of RESERVED_BYTES was 4012 in r1.24 but in r1.26 it got reduced to 4000 after integrating GT.M V6.3-007 (http://tinco.pair.com/bhaskar/gtm/doc/articles/GTM_V6.3-007_Release_Notes.html#GTM-9073). The ydb349 test was originally written for r1.24 with a max limit of 4012 but after r1.26 the max limit needed to be reduced, which is what this commit does.