Disable r124/jnlunxpcterr subtest (maybe temporarily) on Centos Linux for x86_64 as it causes kernel errors in syslog; Fix com/set_gtm_machtype.csh to set linux distrib even if /etc/os-release has double-quotes in ID (e.g. `"centos"` -> `centos`)
Note that the check in instream.csh can do a `"centos" == $gtm_test_linux_distrib` check only
because `com/set_gtm_machtype.csh` was fixed to strip out any double-quotes in the OS name
from /etc/os-release. This is because a CentOS /etc/os-release contains an `ID="centos"`
whereas an Ubuntu box contains an `ID=ubuntu`. The extra double-quotes from `"centos"` need
to be removed for simple checks of this variable to work like in r124/instream.csh.