Fix rare r126/ydb382 subtest failure due to test timing issue
This subtest failed once in lots of test runs in in-house testing with the following diff.

```
118c118
< %YDB-I-MLKCLEANED, LOCK garbage collection freed 2 lock slots for region DEFAULT
---
> %YDB-I-MLKCLEANED, LOCK garbage collection freed 1 lock slots for region DEFAULT
```

The test expected 2 lock slots (that of ^b and ^c) to be cleaned up. But only 1 slot got
cleaned up.

This is because the test starts a LKE process in the background a few steps before and it
sends a MUPIP STOP to it before starting the next phase of the test. But it does not wait
for the process to terminate before moving on to the next phase of the test. This means
it is possible the database shared memory segment corresponding to the backgrounded LKE
phase of the test might still be active when the next phase of the test starts and that
might result in the backgrounded LKE from the previous stage of the test unintentionally
cleaning up an orphaned lock slot from the current phase of the test (and in turn causing
1 less lock slot to be freed up in the MLKCLEANED message).

This test timing issue is now fixed by waiting for the backgrounded LKE process to terminate
before moving on to the next phase of the test.