Fix v63003/gtm4212 test failure
This addresses a test failure that showed up in in-house testing sometimes depending upon the length of the test directory's path. This test would create directories consisting of 25 a's followed by a '/' and then more a's to test if directories are too long. If that chain of a's and '/'s ended with a '/', the test would fail because the first path that should have generated a YDB-E-FILENAMETOOLONG would instead generate a file already exists error due to the "//{filename}" at the end of the path. The fix is to generate path names with large numbers of a's. This is the same thing the test was doing, when it passed, before the recent change.