Fix r100/objlvlchk subtest failure by bumping maxver from V63001 to V63006
On certain hosts in-house we do not have older versions in the range V62000 thru V63000A
and the r100/objlvlchk subtest failed there because it chose a random version in that
non-existent range. But there is no need for the range to be so limited since the object
file format changed as recently as V63006. Therefore the test has been enhanced to choose
from an older version as high as V63006. This causes the non-existent range to be filled
with some versions on all hosts in-house.