Fix v53003/C9I06002996 subtest failure by bumping maxver from V63000A to V63007
On certain hosts in-house we do not have older versions less than or equal to V63000A
and the v53003/C9I06002996 subtest failed there because it chose a random version in that
non-existent range. But there is no need for the range to be so limited since the minor
db version format changed as recently as V63007. Therefore the test has been enhanced to
choose from an older version as high as V63007. This causes the non-existent range to be filled
with some versions on all hosts in-house.