Exclude on-the-fly trigger upgrade if ydb_test_exclude_trig_upgrade env var is defined/set
On certain hosts in-house we do not have older versions in the range V62000 thru V62002.
A lot of the subtests in the `triggers` test do on-the-fly trigger upgrade test and this
requires exactly those older versions. And because of no existing older versions, these
subtests fail. The fix is to set the env var `ydb_test_exclude_trig_upgrade` at system
startup and changing the test framework to exclude this on-the-fly trigger upgrade if that
env var is set.