[YDB#550] Tests to ensure that nested ydb_tp_s() and ydb_tp_st() calls can rollback without requiring the entire outer transaction to rollback.
This commit adds a new ydb550 subtest to each simpleapi and simplethreadapi. These tests both perform nested transactions which involves a series of set operations. The outer transaction sets subscript "1" of the variable. Then a nested transaction sets subscript "2" of the variable and either rolls back or returns a user-specified error code. Then the outer transaction sets subscript "3". The test then verifies that subscripts "1" and "3" are set but subscript "2" is not set.

The simpleapi version of the test uses 2 processes that run concurrently, each setting its own pair of variables. The simplethreadapi version uses 4 threads, each of which sets its own pair of variables. For the ^UPD variables, the inner transaction is rolled back. For the ^USD variables, the inner transaction returns a user-specified error.

There is also a ydb550b.c file that is run as part of the test. It runs 4 processes on a 10 second timer each doing similar updates on the same global variable. The simplethreadapi version runs 4 threads in each process for a total of 16 different threads running for 10 seconds each. This part of the test ensures that some YDB_TP_RESTARTs are generated and that they are handled correctly. If everything is working as it is supposed to, the ydb550b.c part of this test produces no output.