[YDB#567] Test that a MUPIP INTRPT during a hang does not cause an assert failure
This test uses job to start a process to run hang commands. Then it triggers a MUPIP INTRPT during one such hang which leads to an assert failure on debug builds that do not contain the #567 changes. If the build contains the #567 changes, it runs for about 10 seconds then prints a success message if the total time elapsed is less than twice the sum of all of the hang times.