[YDB#520] Address r130/ydb520 test failure when /bin/sh isn't bash
Part C of the r130/ydb520 test failed on a number of machines in internal testing because an export command resulted in a "bad variable name" error. After investigating, it turns out that /bin/sh is set to dash on those machines which uses a different syntax to set a variable. This has been fixed by running the bash script for part C in bash instead of running it in sh.