Fix v63008/gtm9110 subtest failure on some machines due to CLISTRTOOLONG message text differences
This subtest failed on some boxes because $ydb_dist (the first parameter in the command line $ydb_dist/yottadb -run ...) got expanded to a real path with slightly bigger lengths causing the total lenth to become bigger than the maximum allowed YottaDB command line length (MAX_LINE macro defined in sr_unix/cli.h to be 32767+256). This causes a few bytes from the tail of the 33KiB length M program name in the command line to be truncated and resulted in CLISTRTOOLONG error message displaying a slightly different number than expected.

The fix for this is to reduce the command line length by 30 bytes (reduced from 33000 bytes to 32970 bytes). This lets the entire command line length stay withing the MAX_LINE YottaDB command line length limit (and so no truncation happens) and the CLISTRTOOLONG error message text is the same everywhere.