[YDB#550] Disable assert in simplethreadapi/ydb550 that at least one YDB_TP_RESTART happened for 1-CPU machines
This commit disables, on 1-CPU machines, the assert that checks that at least one YDB_TP_RESTART happened in part B of the simplethreadapi/ydb550 test. This assert failed in our internal testing, likely because part B of the test does not run for long enough to guarantee that a YDB_TP_RESTART will happen on a slow single CPU machine in each of the 16 concurrent threads across 4 processes.

This commit also makes a change to the assert in part B of simplethreadapi/ydb550 that ydb_timer_start_t() returned a status of YDB_OK. If the status is not YDB_OK, it now prints out the status code and fails with a YDB_ASSERT(FALSE). The assert failed in our internal testing on a production build on a 1-CPU machine and the status code was not accessible in the core.