New v63008/gtm9092 subtest to test GTM-9092 in V6.3-008
In V6.3-008 ^%YGBLSTAT provides an extrinsic function that returns TRUE(1) if the process in the region is sharing statisitcs or FALSE(0) if it is not sharing statistics. An empty string will be returned if the pid is invalid or there is no sharing in the specified region. This test uses $$IN^%YGBLSTAT to verify that a 1 or 0 is returned for sharing statistics and an empty string for any errors.