Fix reliable v55000/GTM6813 subtest failure/hang on ARMV7L platform
The `v55000/GTM6813` subtest was failing rather reliably on the ARMV7L platform.
The test jobs off an interrupter process and has a `for` loop (line 62 below) where it keeps checking
if a certain number of interrupts have occurred and certain number of seconds of time have elapsed.

```m
61  ; Exit the loop if at least "minintrptcnt" interrupts have been seen AND at least "minelapsedtime" seconds have elapsed
62  for i=1:1:limit quit:(intrptcnt>minintrptcnt)&(minelapsedtime'>$$^difftime($horolog,starttime))  do adder(i,i*2,i*3,i*4,i*5)
```

If so it does a `quit` out of the for loop. But this `quit` condition never triggered.

I incremented a global inside the `handler` label (which is where any interrupt would transfer control to)
to record every interrupt that happened and I could clearly see hundreds of interrupts happened within a
few seconds. But even though the test kept running for minutes (way above its elapsed time limit of 15 seconds),
it never quit out of the for loop.

From what I could see, the `do adder()` call in the for loop never returned back to the next iteration
of the for loop. And because of this, the `quit` postconditional never got evaluated.

As to why `do adder()` call never returned back, I noticed that the interrupter process was sending
interrupts every 0.0005 seconds (i.e. every 0.5 millisecond on average). I suspected this is lower than
the time taken by the process to handle the interrupt and so the interrupted process ends up permanently
spending its time handling interrupts and never doing its main job of executing the for loop.

To test this out, I bumped the `intrptdelay` variable from `0.0001` seconds to `0.001` seconds. That is
increased it from `0.1` milliseconds to `1` millisecond. With that change, the test reliably passed
whereas it previously reliably hung.

It is not clear why the very small value of delay between interrupts causes problems only on the ARMV7L
platform (and not on ARMV6L and AARCH64). But given 1 millisecond between interrupts is a sufficiently
small inter-interrupt time and that works, I am not as concerned about it.

Therefore went ahead with this change for all platforms.

Also took this opportunity to reduce the elapsed time minimum from 15 seconds to 5 seconds. This way
the test will not keep running unnecessarily if the desired number of interrupts are seen in 5 seconds
itself. I expect at least `x86_64` (and likely `ARMV7L` and `AARCH64` too) to run this test 10 seconds
faster due to this change.