Fix rare v63004/gtm8894 subtest failure due to extra output from "strings" command
In some cases `strings $ydb_dist/libyottadb.so | grep $ZRELDATE` would print an extra output (depending on whether printable ascii text lies adjacent to the location where $ZRELDATE lies in the executable). If we try to do an exact match on the strings output that would fail in the extra output case. The fix is to count how many lines have the $ZRELDATE pattern (with or without the extra output) and verify it is exactly 2 and signal a failure otherwise.