Fix rare v60000/gtm4525b subtest failure (TEST-E-IMPTPSTARTTIMEOUT error on slow machines)
As part of a prior commit (7ed0ca5c0173bfadf93c85115930fee67f0e5e69), the timeout to wait in
`startwait^gtm4525b` had been bumped from 600 seconds to 3600 seconds (1 hour).

But turns out even this is not enough in case imptprust gets randomly chosen as the imptp flavor.
The rust compile alone took 1 hour and 15 minutes on a Pi Zero resulting in a test failure.

Therefore, the timeout is now bumped to 3 hours (slightly more than 2x the time seen in a test failure).