[YDB#568] Fix more r130/ydb568 test failures due to timing
After the previous commit, there were some additional test failures on RHEL7/CentOS machines due to the shell prompt discrepancies involving an old shell prompt being shown followed by a CTRL+G then a new shell prompt. The issue was that sending the `q` to the more process quit the more process and displayed a terminal prompt. Then the `\r` went to the terminal prompt which it printed followed by a new terminal prompt. However, the expect could see either prompt as the `SHELL$` which meant that it could send the `echo 9876543` in between the 2 prompts or, on machines where the prompt was 2 prompts separated by a CTRL+G, it could send it after the CTRL+G but before the second prompt. If it did either of these things, the test would fail. By removing the `\r` from `q\r`, this commit fixes the failure by ensuring that the `expect -exact "SHELL$ "` will see the shell prompt produced by the `send -- "q"` which will then wait for the `echo 9876543` to arrive. The rest of the test then operates as intended.