[YDB#568] Fix r130/ydb568 test failure on newer versions of expect
In in-house testing, the r130/ydb568 test started failing on some systems after a system update, likely due to a new version of expect. After the update, the MUPIP EXTRACT STDOUT section of the test would only produce 1 line starting with `--More^x` rather than 2. This caused the test to fail because it was expecting 2 such lines. This commit fixes the failure by filtering the output to filter out all but the first such line ensuring that the test will pass on all machines regardless of what version of expect they are running.