[YottaDB/DB/YDB#619] Remove unnecessary `call_check_retry!` macro
Originally this was added to be a stable version of [`Try`](https://doc.rust-lang.org/std/ops/trait.Try.html). However, now that it does not handle `YDB_ERR_TPRESTART` specially, imptp.rs can just use the `?` operator directly. So there is no need for the macro.

This is mostly cleanup from YDBTest!1021.