[YDB#551] r132/ydb551 $ZSYSLOG breaks formatting of certain strings passed in
Added tests for case where ZSYSLOG has an ! to ensure that the FAO
directive is properly avoided.

Checks against values that are written into syslog to ensure expected
results

For instance, !AD must appear as !AD in syslog
Likewise Hello world! must also appear as it is
and the ! should not be interpreted as start of
FAO directive