Fix extremely rare triggers/gtm8399 test failure
The triggers/gtm8399 subtest failed in in-house testing because the portion of the test that randomly fuzzes #COUNT and #CYCLE in triggers produced the specific combination of random numbers that caused reduction to be set equal to total, making the second argument to validate() 0. This meant that the part of validate that deletes triggers did not delete any triggers and the line that normally reads `All existing triggers (count = "_newtotal_") deleted` instead read `No matching triggers for deletion`. This resulted in the count variable never getting initialized and the pass being incremented 1 time less than expected, causing the test to fail and the zwrite in the failure output to produce an error, resulting in the error trap doing a zshow and terminating the test early.

This commit fixes this failure by setting the delstr variable equal to `No matching triggers found for deletion` if the second argument to validate is 0. Otherwise the delstr variable is set the same as it was before. This change has been tested and passes in both cases. The newtotal=0 case was tested by hard-coding the exact 1-in-billions output of $RANDOM required to produce a second argument of 0 for the validate call and testing to make sure it passed and that `No matching triggers found for deletion` appeared in the trg_delete.outx file.