Avoid CALLINAFTERXIT error message preceding panic due to MUPIP STOP in impjobgo.go
A test was seen in rundown_argless_1/ipcsnoreset where the expected panic due to a MUPIP
STOP was preceeded with the following message:

 IMPJOBGO: Stopping due to error: %YDB-E-CALLINAFTERXIT, After a ydb_exit(), a process cannot create a valid YottaDB context

This suddenly appearing (on a slower 4 core PI system) is due to YDB#632 which, while
not changing the conditions when CALLINAFTERXIT appears, made it appear sooner (with
less processing) than it previously appeared. This allowed it to show up before the
panic operating in a different thread took the process down. The fix is to check for
this error before putting out this error and DO NOT put the error out if the error
is CALLINAFTERXIT which is something we EXPECT to happen.