Attempt to fix expect/gtm8232 test failure involving timeout on 32 bit ARM machine by doubling timeout
The expect/gtm8232 test failed in internal testing on a 32 bit ARM machine because a 30 second timeout expired. This commit attempts to fix this failure by doubling the timeout to 60 seconds. It is not clear if this new timeout is enough to avoid failures. If it is not, we will double that in a later commit and so on until the failures stop happening.