[V63009][YDB#441] New r132/ydb441 test to verify that Auto-ZLINK works correctly when zlinking a new version of an M program after changing $ZROUTINES
This test verifies that YDB#441/GTM-9114 is fixed. This bug was introduced in V6.3-006 and was fixed by V6.3-009. In rare circumstances, an attempt to zlink an M file that had been updated after changing $ZROUTINES could produce an assert failure (on a debug build) or a GTMCHECK fatal error on a pro build. This bug was present in YottaDB versions r1.26, r1.28 and r1.30.