Weekend test fixes for rollback_A/D9E04002440, v63002/gtm8165, and suppl_inst_A/supplementary
For rollback_A/D9E04002440, the test was not waiting until data was available in the secondary's
journal file that was to be manually corrupted. This change waits till something happens on that
side which can take more than 30 seconds on slow PI systems especially on the ARMV6L systems we
have for testing that are all single core. The symptom seen if this fix is not present is on slow
PI systems, is an assert failure from mupip with the following traceback:
> Thread 1 (Thread 0xb6f31220 (LWP 9341)):
> #0  __pthread_kill (threadid=<optimized out>, signo=3) at ../sysdeps/unix/sysv/linux/pthread_kill.c:56
> #1  0xb5ea2f34 in gtm_dump_core () at /Distrib/YottaDB/V999_R131/sr_unix/gtm_dump_core.c:74
> #2  0xb5ea4864 in gtm_fork_n_core () at /Distrib/YottaDB/V999_R131/sr_unix/gtm_fork_n_core.c:163
> #3  0xb5e750e8 in ch_cond_core () at /Distrib/YottaDB/V999_R131/sr_unix/ch_cond_core.c:80
> #4  0xb5fc5d98 in rts_error_va (csa=0x0, argcnt=7, var=...) at /Distrib/YottaDB/V999_R131/sr_unix/rts_error.c:192
> #5  0xb5fc4f88 in rts_error_csa (csa=0x0, argcnt=7) at /Distrib/YottaDB/V999_R131/sr_unix/rts_error.c:99
> #6  0xb6314df8 in mur_prev (jctl=0x6167020, dskaddr=65960) at /Distrib/YottaDB/V999_R131/sr_port/mur_read_file.c:323
> #7  0xb6318004 in mur_fread_eof (jctl=0x6167020, rctl=0xfd2020) at /Distrib/YottaDB/V999_R131/sr_port/mur_read_file.c:710
> #8  0xb63049a0 in mur_open_files (retry=0) at /Distrib/YottaDB/V999_R131/sr_port/mur_open_files.c:901
> #9  0xb62a3628 in mupip_recover () at /Distrib/YottaDB/V999_R131/sr_port/mupip_recover.c:197
> #10 0xb5e22e3c in mupip_main (argc=8, argv=0xbec85274, envp=0xbec85298) at /Distrib/YottaDB/V999_R131/sr_unix/mupip_main.c:122
> #11 0x00011460 in dlopen_libyottadb (argc=8, argv=0xbec85274, envp=0xbec85298, main_func=0x115c4 "mupip_main") at /Distrib/YottaDB/V999_R131/sr_unix/dlopen_libyottadb.c:152
> #12 0x00010ac4 in main (argc=8, argv=0xbec85274, envp=0xbec85298) at /Distrib/YottaDB/V999_R131/sr_unix/mupip.c:22

For v63002/gtm8165, the issue is that ARMV7L systems are now seeing the same issue that ARMV6L systems
have been seeing for a long time (since commit bdfe72b1 aka MR !436). Note this change also reverts
commit 08b66df1 to increase the timeout from 1 mins to 5 mins as no timeout change would fix this issue
since it is a hang instead.

For suppl_inst_A/supplementary, added setting gtm_test_spanreg to 0 to avoid random mapping issues
that cause the test to fail. The test expects local imptp updates on INST3 to update journal files
in all regions (as it checks for RegSeqno in the journal file headers for Stream 0 and Stream 1 in all
journal files and is included in the reference file). Need to disable random global mappings to ensure
this.