[V63009] Fix r126/ydb440 test failure by updating the outref
After integrating the V6.3-009 changes, the r126/ydb440 test was failing in internal testing. The test prints out a comment stating that the grep of "OC_" should produce an `OC_COMMARG` if the compilation was successful. Since `OC_COMMARG` still shows up in the output, the outref needed to be updated to reflect upstream changes to the assembly code generated.