Re-enable v64003/gtm8916 removing all restrictions on it, modify coredump_filter to include file backed mapped memory
The v64003/gtm8916 subtest has been disabled for all glibc versions > 2.25 (though comment says >=) due
to an issue that was found with glibc 2.25. The lower glibc version in use right now (excepting RH 7 systems)
is 2.27. After several (200+) test runs that seem to indicate this test is now running fine, have re-enabled
it without restriction.

Had an issue where the database mapped memory blocks in an MM database were not in a core causing considerable
difficulty in figuring out what went wrong. We are adding the two file-backed-mappend-memory dump flags to
coredump_filter in the test system (in submit_test.csh) so this memory gets dumped from now on.